refactor(workflows): extract instance/step routes into lib/routes/instances (C2b)
Phase C2b of the workflows/index.ts split. Pulls the agent-facing runtime
route group out of populateWorkflowRoutes into a typed defineRoute[] module.

- lib/trigger-dispatch.ts — the fire-and-forget triggerDispatch helper,
  shared by the step routes, the gate routes (still in index for now), and
  the activate-body exec tools.
- lib/routes/instances.ts — exports `instanceRoutes`: /steps/:taskId GET,
  /steps/:taskId/complete POST, /instances GET, /instances/:taskId GET,
  /instances/start POST.

populateWorkflowRoutes is now gate-routes-only; definePlugin spreads
`[...definitionRoutes, ...instanceRoutes, ...workflowRoutes]`. index.ts drops
its triggerDispatch definition (imported now) and the now-unused
getWorkflowPluginContext import (its only reader, the start route, moved).

Pure behavior-preserving relocation — route objects byte-identical.

Verified: typecheck + eslint + workflows suite (420/0) + full suite (5124/0)
+ binary build + isolated-home boot smoke (GET /instances → 200, GET /steps/x
→ 404, POST /instances/start with a bad workflow → 400, /definitions → 200).
